test(WP-4.10): adversarial fuzz — Forge + proptest properties
Five Forge fuzz tests (256 runs each) and five Rust proptest properties
(256 cases each) probe the pure helper extracted at WP-4.8 from
both sides. Same logic, two independent fuzzers, mutual cross-check.

Properties exercised:
  1. Self-mentor short-circuits everything else.
  2. Determinism: same inputs → same output, every time.
  3. OK requires strict mentor advantage (mAcc ≤ eAcc never OK).
  4. Capacity is a hard wall (load ≥ cap always rejects).
  5. Trust-floor monotonicity (raising the floor never opens admissions).
  6. Totality: no panic over the full input space (Rust side).

The fuzzer found two real bugs:

  - Arithmetic overflow in `_validatePairing` (Solidity): the gap
    check `mentorAcc <= menteeAcc + gap` panicked when the sum
    exceeded uint32.max. Fix: cast both addends to uint256 before
    summing. Q16.16 inputs are conventionally bounded but the helper
    must remain panic-free over the full uint32 domain. The Rust
    mirror was already overflow-safe (used u64 for the sum).

  - A flawed symmetry assertion in the original fuzz harness:
    I had claimed (A, B) gap-failure implied (B, A) non-OK, which
    is wrong because the matcher is intentionally asymmetric —
    if A is much weaker than B, (B mentors A) is a valid pair.
    Replaced with the correct directional invariant.

Tests: Full Forge 1341 PASSING (+5). Mentor Rust 26 PASSING (+5).
Tripwires green.

+++ contracts/src/MentorMatcher.sol
@@ -225,7 +225,15 @@ contract MentorMatcher {
         if (mentorAcc < trustFloor_) {
             return PairingValidity.MENTOR_BELOW_TRUST_FLOOR;
         }
-        if (mentorAcc <= menteeAcc + minAccuracyGap_) {
+        // Cast to uint256 so a near-uint32.max menteeAcc + gap does
+        // NOT silently wrap. Found by WP-4.10 fuzzer: with Q16.16
+        // semantics inputs SHOULD be ≤ 65536, but the helper can't
+        // assume well-formedness and must remain panic-free over
+        // the full uint32 domain.
+        if (
+            uint256(mentorAcc)
+                <= uint256(menteeAcc) + uint256(minAccuracyGap_)
+        ) {
             return PairingValidity.ACCURACY_GAP_TOO_SMALL;
         }
         if (mentorLoad_ >= mentorCap_) {
